[IMP] Moved method is_authentication_oidc into namespace blueprints
