+++ source/app/blueprints/pages/login/login_routes.py
@@ -50,7 +50,7 @@ from app.models.cases import Cases
 
 from app.util import is_authentication_ldap
 from app.blueprints.responses import response_error
-from app.util import is_authentication_oidc
+from app.blueprints.access_controls import is_authentication_oidc
 from app.datamgmt.manage.manage_users_db import get_active_user_by_login, get_user
 from app.datamgmt.manage.manage_users_db import create_user
